wiped style sheets, added component and styles folders, refactored App.test.js, wiped index.html
