+++ src/App.js
@@ -1,13 +1,12 @@
-import React from 'react';
-import logo from './logo.svg';
-import './App.css';
+import React from 'react'
+import './styles/App.css'
 
 function App() {
   return (
-    <div className="App">
+    <div className='App'>
       <h1>Pintereach</h1>
     </div>
-  );
+  )
 }
 
-export default App;
+export default App
